% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{methods}
\name{annotateCircs}
\alias{annotateCircs}
\alias{annotateCircs,RangedSummarizedExperiment-method}
\title{Load and annotate a list of circRNA candidates}
\usage{
annotateCircs(se, annot.list, assembly = c("hg19", "hg38", "mm10", "rn5",
  "rn6", "dm6", "WBcel235"), fixCoordIndexing = TRUE, ...)

\S4method{annotateCircs}{RangedSummarizedExperiment}(se, annot.list,
  assembly = c("hg19", "hg38", "mm10", "rn5", "rn6", "dm6", "WBcel235"),
  fixCoordIndexing = TRUE, ...)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{annot.list}{list of relevant genomic features generated using
\code{loadAnnotation()}}

\item{assembly}{what genome assembly the input data are coming from}

\item{fixCoordIndexing}{check and fix genomic coordinate indexing?}

\item{...}{other arguments}
}
\description{
Loads a list of splice junctions detected using \code{find_circ.py}
(Memczak et al. 2013; www.circbase.org), applies quality filters,
calculates circular-to-linear ratios, and extends the input with
genomic features
}
