% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateCircs}
\alias{annotateCircs}
\title{Load and annotate a list of circRNA candidates}
\usage{
annotateCircs(circs.bed, annot.list, assembly = c("hg19", "hg38", "mm10",
  "rn5", "dm6"))
}
\arguments{
\item{circs.bed}{a list of splice junctions (linear and circular), generated using \code{find_circ.py}.}

\item{annot.list}{list of relevant genomic features generated using \code{loadAnnotation()}}

\item{assembly}{what genome assembly the input data are coming from}
}
\description{
Loads a list of splice junctions detected using \code{find_circ.py}
(Memczak et al. 2013; www.circbase.org), applies quality filters,
calculates circular-to-linear ratios, and extends the input with
genomic features
}

