% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateFlanks}
\alias{annotateFlanks}
\title{title}
\usage{
annotateFlanks(se, annot.list)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{annot.list}{an annotation list, as returned by \code{loadAnnotation()}}
}
\description{
description
}
\details{
details
}
