% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateHostGenes}
\alias{annotateHostGenes}
\title{title}
\usage{
annotateHostGenes(se, genes.gr)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{genes.gr}{a GenomicRanges object with gene models}
}
\description{
description
}
\details{
details
}
