% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{AnnotateRanges}
\alias{AnnotateRanges}
\title{title}
\usage{
AnnotateRanges(r1, l, ignore.strand = FALSE, type = "precedence",
  null.fact = "None", collapse.char = ":")
}
\arguments{
\item{circs}{}
}
\description{
description
annotates the ranges with the corresponding list
}
\details{
details
}

