% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\docType{methods}
\name{annotPie}
\alias{annotPie}
\alias{annotPie,RangedSummarizedExperiment-method}
\title{A piechart of gene features input circRNAs are spliced from}
\usage{
annotPie(se, other.threshold = 0.02, ...)

\S4method{annotPie}{RangedSummarizedExperiment}(se, other.threshold = 0.02,
  ...)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{other.threshold}{a minimum number of candidates feature should
have to be present in the pie-chart. Can be expressed
as fraction, or raw number.}

\item{...}{other arguments}
}
\value{
ggplot2 pie-chart
}
\description{
Describes proportions of circRNAs coming from particular gene features
(coding sequence, UTRs, introns, intergenic regions, ...). Low-frequency
features can be collapsed to "other".
}
