% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{annotPie}
\alias{annotPie}
\title{A piechart of gene features input circRNAs are spliced from}
\usage{
annotPie(se, other.threshold = 0.02, ...)
}
\arguments{
\item{other.threshold}{a minimum number of candidates feature should
have to be present in the pie-chart. Can be expressed
as fraction, or raw number.}

\item{circs}{a list of circRNA candidates, loaded and annotated using \code{annotateCircs()}}
}
\value{
ggplot2 pie-chart
}
\description{
Describes proportions of circRNAs coming from particular gene features
(coding sequence, UTRs, introns, intergenic regions, ...). Low-frequency
features can be collapsed to "other".
}

