% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_1}
\alias{anova_1}
\title{anova_1}
\format{An object of class \code{anova} (inherits from \code{data.frame}) with 2 rows and 5 columns.}
\usage{
anova_1
}
\description{
Model of class anova
}
\keyword{datasets}
