% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_3}
\alias{anova_3}
\title{anova_3}
\format{An object of class \code{anova} (inherits from \code{data.frame}) with 6 rows and 5 columns.}
\usage{
anova_3
}
\description{
Model of class anova
}
\keyword{datasets}
