% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_4}
\alias{anova_4}
\title{anova_4}
\format{
An object of class \code{anova} (inherits from \code{data.frame}) with 3 rows and 6 columns.
}
\usage{
anova_4
}
\description{
Model of class anova
}
\keyword{datasets}
