% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_lmerMod_0}
\alias{anova_lmerMod_0}
\title{anova_lmerMod_0}
\format{An object of class \code{anova} (inherits from \code{data.frame}) with 0 rows and 4 columns.}
\usage{
anova_lmerMod_0
}
\description{
Model of class anova of merMod
}
\keyword{datasets}
