% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_lmerMod_1}
\alias{anova_lmerMod_1}
\title{anova_lmerMod_1}
\format{
An object of class \code{anova} (inherits from \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
anova_lmerMod_1
}
\description{
Model of class anova of merMod
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
anova(lmer(wt ~ cyl + (1 | gear), data = mtcars))
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
