% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_lmerMod_3}
\alias{anova_lmerMod_3}
\title{anova_lmerMod_3}
\format{An object of class \code{anova} (inherits from \code{data.frame}) with 3 rows and 4 columns.}
\usage{
anova_lmerMod_3
}
\description{
Model of class anova of merMod
}
\examples{
anova(lmer(wt ~ drat * cyl + (1 | gear), data = mtcars))
}
\keyword{datasets}
