% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_lmerMod_6}
\alias{anova_lmerMod_6}
\title{anova_lmerMod_6}
\format{
An object of class \code{anova} (inherits from \code{data.frame}) with 3 rows and 8 columns.
}
\usage{
anova_lmerMod_6
}
\description{
Model of class anova of merMod
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
anova(lmerMod_0, lmerMod_1, lmerMod_2)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
