% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aov_2}
\alias{aov_2}
\title{aov_2}
\format{An object of class \code{aov} (inherits from \code{lm}) of length 13.}
\usage{
aov_2
}
\description{
Model of class aov
}
\examples{
data <- iris
data$Cat1 <- rep(c("X", "X", "Y"), length.out = nrow(data))
data$Cat2 <- rep(c("A", "B"), length.out = nrow(data))

aov(Sepal.Length ~ Species * Cat1 * Cat2, data = data)
}
\keyword{datasets}
