% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aovlist_1}
\alias{aovlist_1}
\title{aovlist_1}
\format{
An object of class \code{aovlist} (inherits from \code{listof}) of length 3.
}
\usage{
aovlist_1
}
\description{
Model of class aovlist
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aov(wt ~ cyl + Error(gear), data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
