% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aovlist_2}
\alias{aovlist_2}
\title{aovlist_2}
\format{An object of class \code{aovlist} (inherits from \code{listof}) of length 3.}
\usage{
aovlist_2
}
\description{
Model of class aovlist
}
\keyword{datasets}
