% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aovlist_3}
\alias{aovlist_3}
\title{aovlist_3}
\format{An object of class \code{aovlist} (inherits from \code{listof}) of length 3.}
\usage{
aovlist_3
}
\description{
Model of class aovlist
}
\keyword{datasets}
