% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\docType{methods}
\name{bedTracks}
\alias{bedTracks}
\alias{bedTracks,RangedSummarizedExperiment-method}
\title{bedTracks}
\usage{
bedTracks(
  se,
  score = "circ.uniq",
  min.score = unlist(ifelse(is.null(score), list(NULL), list(1))),
  max.score = unlist(ifelse(is.null(score), list(NULL), list(1000)))
)

\S4method{bedTracks}{RangedSummarizedExperiment}(
  se,
  score = "circ.uniq",
  min.score = unlist(ifelse(is.null(score), list(NULL), list(1))),
  max.score = unlist(ifelse(is.null(score), list(NULL), list(1000)))
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object with circRNA information}

\item{score}{\code{character} vector (only first element will be used) naming
the assay to use as \code{BED} score (\code{NULL} to omit BED
score and use `.` as a placeholder instead)}

\item{min.score}{\code{numeric} vector (only first element will be used) with
minimal score a circRNA must have to be included in the
\code{BED} output for a given sample (\code{NULL} to include
all circRNAs for all samples, even if they were not detected
in that sample)}

\item{max.score}{\code{numeric} vector (only first element will be used) with
maximal score to be used in \code{BED} output (higher score
will be truncated) (\code{NULL} to keep scores unlimited,}
}
\value{
returns a GRangesList object with BED6 circRNA data per sample
}
\description{
Export circRNA information from given \code{SummarizedExperiment} object to
\code{BED6}-style \code{GRangesList}.
This is useful to generate genome browser tracks showing circRNA candidates
per sample.
}
