% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betareg_1}
\alias{betareg_1}
\title{betareg_1}
\format{An object of class \code{betareg} of length 29.}
\usage{
betareg_1
}
\description{
Model of class betareg
}
\examples{
betareg(yield ~ batch + temp, data = GasolineYield)
}
\keyword{datasets}
