% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betareg_2}
\alias{betareg_2}
\title{betareg_2}
\format{
An object of class \code{betareg} of length 29.
}
\usage{
betareg_2
}
\description{
Model of class betareg
}
\examples{
betareg(I(food/income) ~ income + persons, data = FoodExpenditure)
}
\keyword{datasets}
