% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_2}
\alias{brms_2}
\title{brms_2}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_2
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
dat <- read.table(header = TRUE, text = "
  n r r/n group treat c2 c1 w
  62 3 0.048387097 1 0 0.1438 1.941115288 1.941115288
  96 1 0.010416667 1 0 0.237 1.186583128 1.186583128
  17 0 0 0 0 0.2774 1.159882668 3.159882668
  41 2 0.048780488 1 0 0.2774 1.159882668 3.159882668
  212 170 0.801886792 0 0 0.2093 1.133397521 1.133397521
  143 21 0.146853147 1 1 0.1206 1.128993008 1.128993008
  143 0 0 1 1 0.1707 1.128993008 2.128993008
  143 33 0.230769231 0 1 0.0699 1.128993008 1.128993008
  73 62 1.260273973 0 1 0.1351 1.121927228 1.121927228
  73 17 0.232876712 0 1 0.1206 1.121927228 1.121927228")
dat$treat <- as.factor(dat$treat)

brms::brm(
  r | trials(n) ~ treat * c2,
  data = dat,
  family = binomial(link = logit),
  chains = 1, iter = 500
)}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
