% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_3}
\alias{brms_3}
\title{brms_3}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_3
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
x <- sample(c(1:3), 1000, replace = TRUE)
w <- runif(1000, min = 0.3, max = 3)
y <- rnorm(1000, mean = 0, sd = 1)
data <- data.frame(x, w, y)
data$x <- factor(data$x)

brm(x | weights(w) ~ y, data = data, family = categorical)}
}
\keyword{datasets}
