% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_4bf_1}
\alias{brms_4bf_1}
\title{brms_4bf_1}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_4bf_1
}
\description{
Example of class brmsfit with \code{save_all_pars = TRUE}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brm(Sepal.Length ~ 1, data = iris, save_all_pars = TRUE)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
