% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_4bf_3}
\alias{brms_4bf_3}
\title{brms_4bf_3}
\format{An object of class \code{brmsfit} of length 21.}
\usage{
brms_4bf_3
}
\description{
Example of class brmsfit with \code{save_all_pars = TRUE}.
}
\examples{
\dontrun{
brm(Sepal.Length ~ Petal.Length, data = iris, save_all_pars = TRUE)
}

}
\keyword{datasets}
