% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_5}
\alias{brms_5}
\title{brms_5}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_5
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontrun{
set.seed(333)
brm(mpg ~ factor(cyl), data = mtcars)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
