% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_aterm_1}
\alias{brms_aterm_1}
\title{brms_aterm_1}
\format{
An object of class \code{brmsfit} of length 22.
}
\usage{
brms_aterm_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
brm(am | trials(1) ~ hp,
  data = mtcars, family = binomial(),
  chains = 2, iter = 200,
  backend = "cmdstanr", cores = 2
)}
}
\keyword{datasets}
