% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_aterm_3}
\alias{brms_aterm_3}
\title{brms_aterm_3}
\format{
An object of class \code{brmsfit} of length 22.
}
\usage{
brms_aterm_3
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
data(kidney)
brm(time | cens(censored) ~ age * sex + disease + (1|patient),
    data = kidney, family = lognormal(),
    chains = 2, iter = 200,
    backend = "cmdstanr", cores = 2)}
}
\keyword{datasets}
