% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_beta_1}
\alias{brms_beta_1}
\title{brms_beta_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_beta_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(FoodExpenditure, package = "betareg")
m <- brm(
  I(food / income) ~ income + (1 | persons),
  data = FoodExpenditure,
  family = Beta()
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
