% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_bf_1}
\alias{brms_bf_1}
\title{brms_bf_1}
\format{
An object of class \code{brmsfit} of length 22.
}
\usage{
brms_bf_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brm(
  mpg ~ wt,
  data = mtcars,
  prior = set_prior("normal(0, 1)", class = "b"),
  refresh = 0,
  iter = 200,
  chains = 2
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
