% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_categorical_1}
\alias{brms_categorical_1}
\title{brms_categorical_1}
\format{
An object of class \code{brmsfit} of length 16.
}
\usage{
brms_categorical_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
data(mtcars)
brms::brm(gear ~ mpg, data = mtcars, family = categorical())
}
}
\keyword{datasets}
