% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_categorical_1_wt}
\alias{brms_categorical_1_wt}
\title{brms_categorical_1_wt}
\format{
An object of class \code{brmsfit} of length 16.
}
\usage{
brms_categorical_1_wt
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
data(mtcars)
brms::brm(gear | weights(wt) ~ mpg, data = mtcars, family = categorical())
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
