% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_categorical_2_num}
\alias{brms_categorical_2_num}
\title{brms_categorical_2_num}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_categorical_2_num
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
brm(
  Species ~ Sepal.Width,
  data = iris,
  family = categorical(refcat = "setosa"),
  refresh = 0,
  backend = "cmdstanr",
  algorithm = "pathfinder"
)
}
}
\keyword{datasets}
