% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_chocomini_1}
\alias{brms_chocomini_1}
\title{brms_chocomini_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_chocomini_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
# Define formula and priors
f <- bf(Real ~ 0 + Intercept + (0 + Intercept|Participant),
        delta ~ 0 + Intercept + (0 + Intercept|Participant),
        phi ~ 0 + Intercept,
        k ~ 0 + Intercept + (0 + Intercept|Participant),
        family = chocomini)

# Define priors (brms::default_prior(f_mini, data=df))
priors <- c(
  prior("normal(0, 0.3)", class = "b", coef = "Intercept"),
  prior("normal(0, 0.3)", class = "b", coef = "Intercept", dpar = "delta"),
  prior("normal(3, 0.5)", class = "b", coef = "Intercept", dpar = "phi"),
  prior("normal(3, 0.5)", class = "b", coef = "Intercept", dpar = "k")
) |> brms::validate_prior(formula = f, data = df)

# Fit the model
brms_chocomini_1 <- brm(
  formula = f,
  data = df,
  prior = priors,
  family = chocomini,
  stanvars = stanvars_chocomini,
  chains = 1,
  iter = 500,
  backend = "cmdstanr"
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
