% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_chocomini_2}
\alias{brms_chocomini_2}
\title{brms_chocomini_2}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_chocomini_2
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontrun{
library(cogmod)
library(brms)

data <- data.frame(
  y = rchoco(
    2000,
    p = 0.5, # Named mu in brms
    confright = 0.8,
    confleft = 0.3
  ),
  cond = rep(c("A", "B"), each = 1000)
)

brm(
  bf(
    y ~ cond, # mu
    confright ~ 1,
    confleft ~ 1,
    precright ~ 1,
    precleft ~ 1,
    pex ~ 1,
    bex ~ 1,
    pmid ~ 1
  ),
  data = data,
  family = choco(),
  stanvars = choco_stanvars(),
  init = 0,
  iter = 500,
  backend = "cmdstanr"
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
