% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_corr_re1}
\alias{brms_corr_re1}
\title{brms_corr_re1}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_corr_re1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
df <- data.frame(x = factor(rep(c('A', 'B'), each = 100)),
                 y = c(rnorm(100, mean = 0, sd = 1),
                       rnorm(100, mean = 10, sd = 5)),
                 id = factor(rep(1:100, 2)))

brm(
  bf(y ~ x + (1 | i | id), sigma ~ x + (1 | i | id)),
  data = df,
  iter = 500,
  cores = 4
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
