% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_intercept_1}
\alias{brms_intercept_1}
\title{brms_intercept_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_intercept_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontrun{
set.seed(123)
df <- data.frame(response = rnorm(100), day = rep(paste0('D', 1:5), each = 20))
brm(
  response ~ 1 + (1 | day),
  data = df,
  seed = 123
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
