% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_ipw_1}
\alias{brms_ipw_1}
\title{brms_ipw_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_ipw_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
brms::brm(
  QoL | weights(ipw) ~ treatment * time + treatment * education +
    hospital + age + phq4 + (1 | ID) ,
  data = datawizard::qolcancer
)
}
}
\keyword{datasets}
