% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_lf_1}
\alias{brms_lf_1}
\title{brms_lf_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_lf_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
brms::brm(
  bf(carb ~ gear * vs) + lf(disc ~ 0 + mo(cyl)),
  data = mtcars,
  family = cumulative("probit")
)
}
}
\keyword{datasets}
