% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_linear_1}
\alias{brms_linear_1}
\title{brms_linear_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_linear_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(efc, package = "ggeffects")
efc <- datawizard::to_factor(efc, c("c161sex", "c172code", "e16sex", "e42dep"))
levels(efc$c172code) <- c("low", "mid", "high")
brm(barthtot ~ e16sex + c161sex + c172code * c160age + c12hour + e42dep, data = efc)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
