% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_meta_1}
\alias{brms_meta_1}
\title{brms_meta_1}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_meta_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
# Data from
# https://github.com/MathiasHarrer/Doing-Meta-Analysis-in-R/blob/master/_data/Meta_Analysis_Data.RData
set.seed(123)
priors <- c(prior(normal(0,1), class = Intercept),
            prior(cauchy(0,0.5), class = sd))

brm(TE|se(seTE) ~ 1 + (1|Author),
   data = Meta_Analysis_Data,
   prior = priors,
   iter = 4000)
}
}
\keyword{datasets}
