% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_miss_1}
\alias{brms_miss_1}
\title{brms_miss_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_miss_1
}
\description{
brms_miss_1
}
\examples{
\dontrun{
dat <- read.csv("https://vincentarelbundock.github.io/Rdatasets/csv/carData/TitanicSurvival.csv")
dat$survived <- ifelse(dat$survived == "yes", 1, 0)
dat$woman <- ifelse(dat$sex == "female", 1, 0) |> as.factor()
sum(is.na(dat$age))  # <-- 263 missing values

brm(
  bf(
    survived ~ woman * mi(age) + passengerClass,
    family = bernoulli(link = "logit")
  ) +
    bf(age | mi() ~ passengerClass + woman) +
    set_rescor(FALSE),
  data = dat,
  backend = "cmdstanr",
  cores = 4
)
}
}
\keyword{datasets}
