% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_1}
\alias{brms_mixed_1}
\title{brms_mixed_1}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mixed_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
brms::brm(mpg ~ wt + (1 | cyl) + (1 + wt | gear), data = mtcars)
}
}
\keyword{datasets}
