% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_10}
\alias{brms_mixed_10}
\title{brms_mixed_10}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mixed_10
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
data(iris)
iris$Group <- as.factor(rep(c("G1", "G2", "G3"), each = 50))
set.seed(123)
brms::brm(
  Sepal.Width ~ Petal.Width + (Petal.Width | Group),
  data = iris,
  refresh = 0
)
}
}
\keyword{datasets}
