% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_2}
\alias{brms_mixed_2}
\title{brms_mixed_2}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mixed_2
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
brms::brm(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
