% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_3}
\alias{brms_mixed_3}
\title{brms_mixed_3}
\format{An object of class \code{brmsfit} of length 20.}
\usage{
brms_mixed_3
}
\description{
Model of class brms
}
\examples{
\dontrun{
set.seed(123)
sleepstudy$grp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy <- sleepstudy \%>\%
  group_by(grp) \%>\%
  mutate(subgrp = sample(1:15, size = n(), replace = TRUE))

brms::brm(Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject), data = sleepstudy)
}
}
\keyword{datasets}
