% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_4}
\alias{brms_mixed_4}
\title{brms_mixed_4}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mixed_4
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(333)
brms::brm(Petal.Length ~ Petal.Width + (1 | Species), data = iris)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
