% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_5}
\alias{brms_mixed_5}
\title{brms_mixed_5}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_mixed_5
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
library(lme4)
library(brms)
data(sleepstudy)
set.seed(123)
sleepstudy$cat <- as.factor(sample(1:5, size = 180, replace = TRUE))
sleepstudy$Reaction_d <-
  ifelse(sleepstudy$Reaction < median(sleepstudy$Reaction), 0, 1)

brms::brm(
  Reaction_d ~ Days + cat + (1 | Subject),
  data = sleepstudy,
  family = bernoulli()
)
}
}
\keyword{datasets}
