% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_6}
\alias{brms_mixed_6}
\title{brms_mixed_6}
\format{An object of class \code{brmsfit} of length 21.}
\usage{
brms_mixed_6
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
bprior1 <- prior(student_t(5, 0, 10), class = b) + prior(cauchy(0, 2), class = sd)
brms::brm(
  count ~ Age + Base * Trt + (1 | patient),
  data = epilepsy,
  family = poisson(),
  prior = bprior1,
  chains = 1,
  iter = 500
)
}
}
\keyword{datasets}
