% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixed_9}
\alias{brms_mixed_9}
\title{brms_mixed_9}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mixed_9
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brms::brm(mpg ~ hp + (1 | cyl), data = mtcars, seed = 123)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
