% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mixture_1}
\alias{brms_mixture_1}
\title{brms_mixture_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mixture_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brm(
  formula = QoL ~ time + hospital + education + age + (1 + time | ID),
  data = qol_cancer, # Your prepared dataset
  family = mixture(gaussian, nmix = 3), # Gaussian distribution for outcome, K_classes components
  chains = 4, # Number of MCMC chains
  iter = 1000, # Total iterations per chain (includes warmup)
  cores = 4,  # Number of CPU cores to use
  seed = 1234 # For reproducibility
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
