% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mm_1}
\alias{brms_mm_1}
\title{brms_mm_1}
\format{An object of class \code{brmsfit} of length 21.}
\usage{
brms_mm_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
dat <- data.frame(
  sid = c(1:300),
  t1id = sample(200:215, 300, replace = TRUE),
  t2id = sample(200:215, 300, replace = TRUE),
  y = rnorm(300, mean = 85, sd = 8),
  x = rnorm(300, mean = 50, sd = 10)
)

brm(y ~ x + (1 | mm(t1id, t2id)), data = dat)
}
}
\keyword{datasets}
