% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mm_2}
\alias{brms_mm_2}
\title{brms_mm_2}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mm_2
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
dat <- data.frame(
  y = rnorm(100), x1 = rnorm(100), x2 = rnorm(100),
  g1 = sample(1:10, 100, TRUE), g2 = sample(1:10, 100, TRUE)
)

# multi-membership model with level specific covariate values
dat$xc <- (dat$x1 + dat$x2) / 2
brm(y ~ xc + (1 + mmc(x1, x2) | mm(g1, g2)), data = dat)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
