% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mm_3}
\alias{brms_mm_3}
\title{brms_mm_3}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mm_3
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
data(mzcars)
mtcars$w <- 1
brm(
  mpg ~ 1 + hp + (1 | cyl) +
    (1 | carb:am:hp) +
    (1 | mm(carb, am, weights = cbind(w, w), scale = FALSE)) +
    (1 | mm(carb:cyl, am:cyl, weights = cbind(w, w), scale = FALSE)),
  data = mtcars, family = gaussian, chains = 4,
  iter = 100, control = list(adapt_delta = 0.99, max_treedepth = 15),
  cores = 4, seed = 12345,
  backend = "cmdstanr"
)
}
}
\keyword{datasets}
