% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mm_4}
\alias{brms_mm_4}
\title{brms_mm_4}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mm_4
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
dat <- data.frame(
  s1 = sample(200:215, 300, replace = TRUE),
  s2 = sample(200:215, 300, replace = TRUE),
  y = rnorm(300, mean = 85, sd = 8),
  x = rnorm(300, mean = 50, sd = 10)
)

brm(y ~ x + (1 | mm(s1, s2)), data = dat)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
