% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mo2}
\alias{brms_mo2}
\title{brms_mo2}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_mo2
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
brm(mpg ~ mo(gear), data = mtcars, refresh = 0)
}
}
\keyword{datasets}
