% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mv_2}
\alias{brms_mv_2}
\title{brms_mv_2}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mv_2
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
f1 <- bf(mpg ~ wt + disp + cyl + hp + (1 |CAR| gear))
f2 <- bf(wt ~ disp + cyl + hp + (1 |CAR| gear))
brms::brm(f1 + f2 + set_rescor(FALSE), data = mtcars)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
