% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mv_3}
\alias{brms_mv_3}
\title{brms_mv_3}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_mv_3
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(epilepsy)
set.seed(123)
epilepsy$visit <- as.numeric(epilepsy$visit)
epilepsy$Base2 <- sample(epilepsy$Base, nrow(epilepsy), replace = TRUE)
f1 <- bf(Base ~ zAge + count + (1 |ID| patient))
f2 <- bf(Base2 ~ zAge + Trt + (1 |ID| patient))
brms::brm(f1 + f2 + set_rescor(FALSE), data = epilepsy)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
