% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mv_4}
\alias{brms_mv_4}
\title{brms_mv_4}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_mv_4
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
f1 <- bf(Sepal.Length ~ Petal.Length + Sepal.Width + Species)
f2 <- bf(Sepal.Width ~ Species)
brms::brm(f1 + f2 + set_rescor(FALSE), data = iris, chains = 1, iter = 500)
}
}
\keyword{datasets}
