% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mv_5}
\alias{brms_mv_5}
\title{brms_mv_5}
\format{An object of class \code{brmsfit} of length 21.}
\usage{
brms_mv_5
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
bf1 <- bf(count ~ child + camper + (1 | persons), zi ~ camper + (1 | persons))
bf2 <- bf(count2 ~ child + livebait + (1 | persons), zi ~ child + (1 | persons))
brms::brm(bf1 + bf2, data = zinb, family = zero_inflated_poisson(), chains = 1, iter = 500)
}
}
\keyword{datasets}
