% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_mv_6}
\alias{brms_mv_6}
\title{brms_mv_6}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_mv_6
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(jobs, package = "mediation")
f1 <- bf(job_seek ~ treat + econ_hard + sex + age)
f2 <- bf(depress2 ~ treat + job_seek + econ_hard + sex + age)
brm(f1 + f2 + set_rescor(FALSE), data = jobs)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
