% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_ordinal_1}
\alias{brms_ordinal_1}
\title{brms_ordinal_1}
\format{An object of class \code{brmsfit} of length 21.}
\usage{
brms_ordinal_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
data(mtcars)
mtcars$cyl_ord <- as.ordered(mtcars$cyl)
brms::brm(cyl_ord ~ mpg, data = mtcars, family = cumulative())
}
}
\keyword{datasets}
