% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_ordinal_1_wt}
\alias{brms_ordinal_1_wt}
\title{brms_ordinal_1_wt}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_ordinal_1_wt
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
data(mtcars)
mtcars$cyl_ord <- as.ordered(mtcars$cyl)
brms::brm(cyl_ord | weights(wt) ~ mpg, data = mtcars, family = cumulative())
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
