% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_ordinal_2}
\alias{brms_ordinal_2}
\title{brms_ordinal_2}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
brms_ordinal_2
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
data(inhaler)
brm(
  rating ~ treat + period + carry + (1 | subject),
  family = cumulative(),
  data = inhaler
)
}
}
\keyword{datasets}
