% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_sigma_2}
\alias{brms_sigma_2}
\title{brms_sigma_2}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_sigma_2
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
d <- data.frame(
  Condition = rep(c("A", "B", "C"), each=30),
  Participant = as.factor(rep_len(paste0("S", 1:5), 90)),
  Response = rnorm(90)
)
brm(
  bf(
    Response ~ Condition + (Condition | Participant),
    sigma ~ Condition + (Condition | Participant),
    family = exgaussian()
  ),
  data = d,
  iter = 2000,
  chains = 4,
  cores = 4,
  backend = "cmdstanr"
)
}
}
\keyword{datasets}
