% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_sigma_3}
\alias{brms_sigma_3}
\title{brms_sigma_3}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_sigma_3
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(123)
data(iris)
iris$Group <- as.factor(rep(c("G1", "G2", "G3"), each = 50))
brms_sigma_3 <- brms::brm(
  brms::bf(
    Sepal.Width ~ Petal.Width + (Petal.Width | Group),
    sigma ~ Petal.Width + (Petal.Width | Group)
  ),
  data = iris,
  refresh=0
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
