% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_slopes_1}
\alias{brms_slopes_1}
\title{brms_slopes_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_slopes_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
states <- as.data.frame(state.x77)
states$HSGrad <- states$`HS Grad`
brm(
  Income ~ HSGrad + Murder * Illiteracy,
  data = states,
  chains = 2,
  iter = 100,
  cores = 2,
  backend = 'cmdstanr'
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
