% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_smooth_1}
\alias{brms_smooth_1}
\title{brms_smooth_1}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_smooth_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
dat <- mgcv::gamSim(1, n = 200, scale = 2)
brm(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat, iter = 1000)
}
}
\keyword{datasets}
