% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_trunc_1}
\alias{brms_trunc_1}
\title{brms_trunc_1}
\format{
An object of class \code{brmsfit} of length 22.
}
\usage{
brms_trunc_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(epilepsy)
brms_trunc_1 <- brm(
  count | trunc(ub = 104) ~ zBase * Trt,
  data = epilepsy,
  family = poisson(),
  backend = "cmdstanr",
  cores = 4
)}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
