% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_zi_1}
\alias{brms_zi_1}
\title{brms_zi_1}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
brms_zi_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
zinb <- read.csv("http://stats.idre.ucla.edu/stat/data/fish.csv")
set.seed(123)
brms::brm(bf(
    count ~ persons + child + camper,
    zi ~ child + camper
  ),
  data = zinb,
  family = zero_inflated_poisson()
)
}
}
\keyword{datasets}
