% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_zi_2}
\alias{brms_zi_2}
\title{brms_zi_2}
\format{An object of class \code{brmsfit} of length 20.}
\usage{
brms_zi_2
}
\description{
Model of class brms
}
\examples{
\dontrun{
zinb <- read.csv("http://stats.idre.ucla.edu/stat/data/fish.csv")
set.seed(123)
brms::brm(bf(
    count ~ persons + child + camper + (1 | persons),
    zi ~ child + camper + (1 | persons)
  ),
  data = zinb,
  family = zero_inflated_poisson()
)
}
}
\keyword{datasets}
