% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_zi_3}
\alias{brms_zi_3}
\title{brms_zi_3}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_zi_3
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
brms::brm(
  bf(count ~ child + camper + (1 | persons), zi ~ child + camper + (1 | persons)),
  data = zinb,
  family = zero_inflated_poisson(),
  chains = 1,
  iter = 500
)
}
}
\keyword{datasets}
