% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_zi_4}
\alias{brms_zi_4}
\title{brms_zi_4}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_zi_4
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brms::brm(
  bf(
    count ~ child + camper + (1 + xb | persons) + (1 + zg | ID),
    zi ~ child + livebait + (1 + zg + nofish | ID)
  ),
  data = zinb,
  family = zero_inflated_poisson(),
  chains = 1,
  iter = 500
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
