% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_zoib_1}
\alias{brms_zoib_1}
\title{brms_zoib_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_zoib_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
df <- data.frame(
  y = cogmod::rbext(1000),
  participant = rep(paste0("S", 1:10), each = 100)
)

f <- bf(
  y ~ 1 + (1 | participant),
  phi ~ 1 + (1 | participant),
  zoi ~ 1 + (1 | participant),
  coi ~ 1 + (1 | participant)
)

brm(f, data = df, family = zero_one_inflated_beta(), init = 0,
    chains = 4, iter = 500, backend = "cmdstanr", refresh = 0)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
