% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTdata}
\alias{BTdata}
\title{BTdata: Morphological Traits of the Blue Tit (Cyanistes caeruleus)}
\format{
A data frame
\enumerate{
\item tarsus:	Tarsus length, z-transformed
\item back:	Colour score of back plummage, z-transformed
\item animal:	Unique identification for each chick (828 levels)
\item dam:	Unique identification for each mother (106 levels)
\item fosternest:	Unique identification for foster nest (104 levels)
\item hatchdate:	Hatching date of chick, z-transformed
\item sex:	Sex of chick (3 levels, Fem = Female, Male = Male, UNK = Unknown)
}
}
\usage{
BTdata
}
\description{
This dataset was collected for the study: "Testing the phenotypic gambit: phenotypic, genetic and environmental correlations of colour" by Hadfield et. al (2007). Data was collected in the spring of 2002 on a population of blue tits at Silwood Park, London, UK. This dataset includes seven variables measured from a cohort of blue tit chicks (n = 828). Blue tit chicks were cross-fostered between nests (i.e. chicks from different families were swapped). Variables include: two morphological traits (tarsus and back); sex; hatching date; the ID of the chick; ID of the chick's mother and the ID of the nest that fostered the chick.
}
\references{
Hadfield, J. D., Nutall, A., Osorio, D., & Owens, I. P. F. (2007). Testing the phenotypic gambit: phenotypic, genetic and environmental correlations of colour. Journal of Evolutionary Biology, 20(2), 549-557.
}
