% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censReg_1}
\alias{censReg_1}
\title{censReg_1}
\format{An object of class \code{censReg} (inherits from \code{maxLik}, \code{maxim}, \code{list}) of length 20.}
\usage{
censReg_1
}
\description{
Model of class censReg
}
\examples{
censReg(
  affairs ~ age + yearsmarried + religiousness + occupation + rating,
  data = Affairs
)
}
\keyword{datasets}
