% Generated by roxygen2 (4.0.1): do not edit by hand
\name{circ_embed}
\alias{circ_embed}
\title{Exact simulation by circulant embedding}
\usage{
circ_embed(x, n = 1000)
}
\arguments{
\item{x}{original time series}

\item{n}{number of simulations (must be a multiple of 2)}
}
\value{
a data.frame with n simulated times series in columns.
}
\description{
Circulant embedding is used to simulate the time series data n
times as time series with the same frequency characteristics like
the original time series (Percival & Constantine, 2006). The
resulting simulated time series are scaled to the mean and
standard deviation of the original time series.
}
\examples{
x <- rnorm(100)
xsim <- circ_embed(x)
}

