% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\docType{methods}
\name{circLinRatio}
\alias{circLinRatio}
\alias{circLinRatio,RangedSummarizedExperiment-method}
\title{circLinRatio}
\usage{
circLinRatio(se, ...)

\S4method{circLinRatio}{RangedSummarizedExperiment}(se, ...)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{...}{other arguments}
}
\value{
returns a data.table
}
\description{
circLinRatio
}
