% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantification.R
\name{circLinRatio}
\alias{circLinRatio}
\title{title}
\usage{
circLinRatio(sites, label.circ = "circ", label.norm = "norm",
  return.readcounts = FALSE)
}
\arguments{
\item{sites}{splice site candidates, as read from find_circ.py output by readCircs}

\item{label.circ}{a string, label used in the name column to distinguish circular from linear splice sites}

\item{label.norm}{a string, label used in the name column to distinguish linear from circular splice sites}
}
\description{
description
}
\details{
details
}

