% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circul.R
\name{circul}
\alias{circul}
\title{Train an unsupervised learning model}
\usage{
circul(x, method = "cmeans", metric = NULL, circulControl = NULL,
  tuneGrid = NULL, modelType = "Classification", tuneLength = 3)
}
\arguments{
\item{x}{a data.frame consisting of measures to be clustered}

\item{method}{a method in the circul model library, character}

\item{metric}{a character representing the performance metric to calculate}

\item{tuneGrid}{a tuning grid of paramater values}

\item{modelType}{the type of model}

\item{tuneLength}{the length of the tuneGrid}

\item{cirControl}{a list of control parameters}
}
\value{
a data.frame with performance metrics
}
\description{
Train an unsupervised learning model
}

