% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circus.R
\docType{package}
\name{ciRcus}
\alias{ciRcus}
\alias{ciRcus-package}
\title{ciRcus: An R package for circRNAs munging}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

