% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clm2_1}
\alias{clm2_1}
\title{clm2_1}
\format{An object of class \code{clm2} of length 25.}
\usage{
clm2_1
}
\description{
Model of class clm2
}
\examples{
clm2_1 <- clm2(rating ~ temp * contact, data = wine)
}
\keyword{datasets}
