% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clm_1}
\alias{clm_1}
\title{clm_1}
\format{
An object of class \code{clm} of length 34.
}
\usage{
clm_1
}
\description{
Model of class clm
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clm_1 <- clm(rating ~ temp * contact, data = wine)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
