% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_github}
\alias{download_data_github}
\title{Download rda files from github}
\usage{
download_data_github(url)
}
\arguments{
\item{url}{URL of the rdata file.}
}
\description{
Download rda files from github
}
\keyword{internal}
