% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{ensg2name}
\alias{ensg2name}
\title{title}
\usage{
ensg2name(ensg, organism, release = "current")
}
\arguments{
\item{ensg}{character vector of ensembl gene ids}

\item{organism}{three-letter organism abbreviation, such as hsa, mmu, rno...}

\item{release}{which ensembl release to use?}
}
\description{
description
}
\details{
details
}
