% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{fukuyama.sugeno}
\alias{fukuyama.sugeno}
\title{Calculate the Fukukama-Sugeno Cluster Metric}
\usage{
fukuyama.sugeno(clres)
}
\arguments{
\item{clres}{results of a cluster algorithm}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate the Fukukama-Sugeno Cluster Metric
}

