% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{fuzzySummary}
\alias{fuzzySummary}
\title{Calculate cluster performance metrics for fuzzy clusters}
\usage{
fuzzySummary(clres, metric = "pe")
}
\arguments{
\item{clres}{results of a cluster algorithm}

\item{metric}{a character representing the abbreviation of the metric, "fs",
"pe", "pc", or "propexp", see details}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate cluster performance metrics for fuzzy clusters
}

