% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gam_zi_1}
\alias{gam_zi_1}
\title{gam_zi_1}
\format{
An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 53.
}
\usage{
gam_zi_1
}
\description{
Model of class gam
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mgcv)
f0 <- function(x) 2 * sin(pi * x)
f1 <- function(x) exp(2 * x)
f2 <- function(x) 0.2 * x ^ 11 * (10 * (1 - x)) ^ 6 + 10 * (10 * x) ^ 3 * (1 - x) ^ 10
n <- 500
set.seed(5)
x0 <- runif(n)
x1 <- runif(n)
x2 <- runif(n)
x3 <- runif(n)

eta1 <- f0(x0) + f1(x1) - 3
p <- binomial()$linkinv(eta1)
y <- as.numeric(runif(n) < p) ## 1 for presence, 0 for absence

ind <- y > 0
eta2 <- f2(x2[ind]) / 3
y[ind] <- rpois(exp(eta2), exp(eta2))

gam(list(y ~ s(x2) + s(x3),  ~ s(x0) + s(x1)), family = ziplss())
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
