% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamlss_1}
\alias{gamlss_1}
\title{gamlss_1}
\format{An object of class \code{gamlss} (inherits from \code{gam}, \code{glm}, \code{lm}) of length 88.}
\usage{
gamlss_1
}
\description{
Model of class gamlss
}
\examples{
data(abdom)
gamlss(
  y ~ pb(x), sigma.formula =  ~ pb(x),
  family = BCT, data = abdom, method = mixed(1, 20)
)
}
\keyword{datasets}
