% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamlss_2}
\alias{gamlss_2}
\title{gamlss_2}
\format{
An object of class \code{gamlss} (inherits from \code{gam}, \code{glm}, \code{lm}) of length 51.
}
\usage{
gamlss_2
}
\description{
Model of class gamlss
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(aids)
gamlss(y ~ x + qrt, data = aids, family = NBI)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
