% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icc.R
\name{get_icc}
\alias{get_icc}
\title{Get intra-class correlation}
\usage{
get_icc(model, var)
}
\arguments{
\item{model}{A model object of class "brmsfit". Model must contains at least one random effect in order to computer an ICC value}

\item{var}{A character string that matches a random effect parameter in the model for which ICC is computed for.}
}
\value{
A data frame containing the variable (var) which ICC is computed for, the posterior mean (Estimate) of ICC value, the standard deviation of the posterior distribution (Est.Error) and the two-sided 95% credible intervals (based on quantiles).
}
\description{
Computes the intra-class coefficient from a "brmsfit" model object using the entire posterior distribution
}
\examples{
#Computes ICC for "ANIMAL", this represents the proportion of variance
#explained by additive genetic variation among individuals (i.e. heritability)
\dontrun{get_icc(model = fit1, var = "ANIMAL")}

#Computes ICC for "MOTHER", this is the proportion of variance
#explained by maternal identity (i.e maternal effects)
\dontrun{get_icc(model = fit1, var = "MOTHER")}
}
\author{
Fonti Kar - fonti.kar@gmail.com
}
