% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getModelInfo}
\alias{getModelInfo}
\title{Get model info from model library}
\usage{
getModelInfo(model = NULL, regex = TRUE, ...)
}
\arguments{
\item{model}{a character vector of the model name}

\item{regex}{logical, should regular expression matching be used?}

\item{...}{additional arguments for the regex}
}
\value{
a list or list of lists of model objects
}
\description{
Get model info from model library
}
\details{
Based on the getModelInfo function in the \code{caret} package
}

