% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_1}
\alias{glm_1}
\title{glm_1}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
glm_1
}
\description{
Model of class glm
}
\details{
\code{glm(formula = vs ~ wt, family = "binomial", data = mtcars)}
}
\keyword{datasets}
