% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_2}
\alias{glm_2}
\title{glm_2}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
glm_2
}
\description{
Model of class glm
}
\examples{
glm(formula = vs ~ wt + cyl, family = "binomial", data = mtcars)
}
\keyword{datasets}
