% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_3}
\alias{glm_3}
\title{glm_3}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
glm_3
}
\description{
Model of class glm
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
glm(formula = vs ~ wt * cyl, family = "binomial", data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
