% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_4}
\alias{glm_4}
\title{glm_4}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
glm_4
}
\description{
Model of class glm
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
glm(  formula = vs ~ wt + cyl,  family = binomial(link = "probit"),  data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
