% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_5}
\alias{glm_5}
\title{glm_5}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
glm_5
}
\description{
Model of class glm
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set.seed(123)
mtcars$count <- rpois(nrow(mtcars), 2)
glm(formula = count ~ wt + cyl, family = "poisson", data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
