% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_nb_1}
\alias{glm_nb_1}
\title{glm_nb_1}
\format{
An object of class \code{negbin} (inherits from \code{glm}, \code{lm}) of length 31.
}
\usage{
glm_nb_1
}
\description{
Model of class glm
}
\examples{
library(MASS)
data(efc, package = "sjmisc")
glm.nb(tot_sc_e ~ neg_c_7 + e42dep + c172code, data = efc)
}
\keyword{datasets}
