% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLMMadaptive_1}
\alias{GLMMadaptive_1}
\title{GLMMadaptive_1}
\format{An object of class \code{MixMod} of length 20.}
\usage{
GLMMadaptive_1
}
\description{
Model of class GLMMadaptive
}
\keyword{datasets}
