% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLMMadaptive_zi_2}
\alias{GLMMadaptive_zi_2}
\title{GLMMadaptive_zi_2}
\format{An object of class \code{MixMod} of length 20.}
\usage{
GLMMadaptive_zi_2
}
\description{
Model of class GLMMadaptive
}
\examples{
\dontrun{
# data prepararion: see "glmmTMB_1"

GLMMadaptive::mixed_model(
  count ~ child + camper,
  random = ~ 1 | persons,
  zi_fixed = ~ child + livebait,
  zi_random = ~ 1 | persons,
  data = fish,
  family = GLMMadaptive::zi.poisson()
)
}
}
\keyword{datasets}
