% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_1}
\alias{glmmTMB_1}
\title{glmmTMB_1}
\format{An object of class \code{glmmTMB} of length 7.}
\usage{
glmmTMB_1
}
\description{
Model of class glmmTMB
}
\examples{
set.seed(123)
fish <- read.csv("https://stats.idre.ucla.edu/stat/data/fish.csv")
fish$nofish <- as.factor(fish$nofish)
fish$livebait <- as.factor(fish$livebait)
fish$camper <- as.factor(fish$camper)
fish$ID <- sample(1:4, nrow(fish), replace = TRUE)

glmmTMB(
  count ~ child + camper + (1 | persons),
  data = fish,
  family = poisson()
)
}
\keyword{datasets}
