% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_3}
\alias{glmmTMB_3}
\title{glmmTMB_3}
\format{An object of class \code{glmmTMB} of length 7.}
\usage{
glmmTMB_3
}
\description{
Model of class glmmTMB
}
\examples{
\dontrun{
glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula =  ~ spp + mined,
  family = nbinom2,
  data = Salamanders
)
}
}
\keyword{datasets}
