% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_5}
\alias{glmmTMB_5}
\title{glmmTMB_5}
\format{
An object of class \code{glmmTMB} of length 7.
}
\usage{
glmmTMB_5
}
\description{
Model of class glmmTMB
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
data(cbpp, package = "lme4")
glmmTMB(
  cbind(incidence, size - incidence) ~ period + (1 | herd),
  data = cbpp,
  family = binomial
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
