% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_spatial_1}
\alias{glmmTMB_spatial_1}
\title{glmmTMB_spatial_1}
\format{An object of class \code{glmmTMB} of length 7.}
\usage{
glmmTMB_spatial_1
}
\description{
Model of class glmmTMB
}
\examples{
\dontrun{
library(geoR)
data(ca20)
dat <- data.frame(
  x = ca20$coords[, 1],
  y = ca20$coords[, 2],
  calcium = ca20$data,
  elevation = ca20$covariate[, 1],
  region = factor(ca20$covariate[, 2])
)
dat$pos <- numFactor(scale(dat$x), scale(dat$y))
dat$ID <- factor(rep(1, nrow(dat)))

glmmTMB(calcium ~ elevation + region + mat(pos + 0 | ID), data = dat)
}
}
\keyword{datasets}
