% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_zi_3}
\alias{glmmTMB_zi_3}
\title{glmmTMB_zi_3}
\format{An object of class \code{glmmTMB} of length 7.}
\usage{
glmmTMB_zi_3
}
\description{
Model of class glmmTMB
}
\examples{
\dontrun{
# data prepararion: see "glmmTMB_1"

glmmTMB(
  count ~ child + camper + (1 | persons),
  ziformula = ~ child + livebait + (1 | persons),
  dispformula = ~xb,
  data = fish,
  family = truncated_poisson()
)
}
}
\keyword{datasets}
