% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gryphon}
\alias{gryphon}
\title{Gryphon: Morphological Traits of the mythical Gryphon}
\format{
A data frame
\enumerate{
\item ANIMAL: Unique identification for each gryphon (1084 levels)
\item MOTHER: Unique identification for each mother (429 levels)
\item BYEAR: Year of Birth
\item SEX:	Sex of gryphon
\item BWT:	Birth weight at fledging
\item TARSUS:	Tarsus length at fledging
}
}
\usage{
gryphon
}
\description{
This dataset was simulated for the paper: "An ecologist's guide to the animal model" by Wilson et. al 2010. Phenotypic traits were simulated over an arbitary pedigree structure using the program 'Pedantics' (Morrissey et al. 2007) to illustrate the uses of the 'animal model'. This dataset includes six variables measured from a cohort of gryphons (n = 1084). Variables include: The ID of each gryphon, the ID of the gryphon's mother, year of birth, sex, birth weight and tarsus length
}
\references{
Wilson, A. J., Reale, D., Clements, M. N., Morrissey, M. M., Postma, E., Walling, C. A., Kruuk, L. E. B & Nussey, D. H. (2010). An ecologist???s guide to the animal model. Journal of Animal Ecology, 79(1), 13-26.

Morrissey, M. B., Wilson, A. J., Pemberton, J. M., & Ferguson, M. M. (2007). A framework for power and sensitivity analyses for quantitative genetic studies of natural populations, and case studies in Soay sheep (Ovis aries). Journal of Evolutionary Biology, 20(6), 2309-2321.
}
\seealso{
gryphonPed
}
