% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gryphonPed}
\alias{gryphonPed}
\title{GryphonPed: Pedigree of the mythical Gryphon}
\format{
A data frame
\enumerate{
\item ANIMAL: Unique identification for each gryphon (1309 levels)
\item FATHER: Unique identification for each father (159 levels)
\item MOTHER: Unique identification for each mother (429 levels)
}
}
\usage{
gryphonPed
}
\description{
This dataset is a pedigree and was simulated for the paper: "An ecologist's guide to the animal model" by Wilson et. al 2010. This dataset includes three variables from a cohort of gryphons (n = 1309). Variables include: The ID of each gryphon, the ID of the gryphon's father and the  the ID of the gryphon's mother
}
\references{
Wilson, A. J., Reale, D., Clements, M. N., Morrissey, M. M., Postma, E., Walling, C. A., Kruuk, L. E. B & Nussey, D. H. (2010). An ecologist???s guide to the animal model. Journal of Animal Ecology, 79(1), 13-26.

Morrissey, M. B., Wilson, A. J., Pemberton, J. M., & Ferguson, M. M. (2007). A framework for power and sensitivity analyses for quantitative genetic studies of natural populations, and case studies in Soay sheep (Ovis aries). Journal of Evolutionary Biology, 20(6), 2309-2321.
}
\seealso{
\code{\link{gryphon}}
}
