% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!
This is an example function named 'hello' which prints 'Hello, world!'.}
\usage{
hello()
}
\description{
Hello, world!
This is an example function named 'hello' which prints 'Hello, world!'.
}
\keyword{internal}
