% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\docType{methods}
\name{histogram}
\alias{histogram}
\alias{histogram,RangedSummarizedExperiment-method}
\title{histogram}
\usage{
histogram(se, binwidth = 0.7, ...)

\S4method{histogram}{RangedSummarizedExperiment}(se, binwidth = 0.7, ...)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{binwidth}{bin width}

\item{...}{other arguments}
}
\value{
returns a ggplot2 object
}
\description{
histogram of SummarizedEXperiment read counts
}
