% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htest_1}
\alias{htest_1}
\title{htest_1}
\format{
An object of class \code{htest} of length 8.
}
\usage{
htest_1
}
\description{
Model of class htest
}
\examples{
cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "spearman")
}
\keyword{datasets}
