% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htest_3}
\alias{htest_3}
\title{htest_3}
\format{An object of class \code{htest} of length 8.}
\usage{
htest_3
}
\description{
Model of class htest
}
\examples{
cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "kendall")
}
\keyword{datasets}
