% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htest_6}
\alias{htest_6}
\title{htest_6}
\format{
An object of class \code{htest} of length 9.
}
\usage{
htest_6
}
\description{
Model of class htest
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
t.test(iris$Sepal.Width, iris$Sepal.Length)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
