% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ivreg_1}
\alias{ivreg_1}
\title{ivreg_1}
\format{
An object of class \code{ivreg} of length 18.
}
\usage{
ivreg_1
}
\description{
Model of class ivreg
}
\examples{
data(CigarettesSW)
CigarettesSW$rprice <- with(CigarettesSW, price / cpi)
CigarettesSW$rincome <- with(CigarettesSW, income / population / cpi)
CigarettesSW$tdiff <- with(CigarettesSW, (taxs - tax) / cpi)

ivreg_1 <- ivreg(
  log(packs) ~ log(rprice) + log(rincome) | log(rincome) + tdiff + I(tax / cpi),
  data = CigarettesSW,
  subset = year == "1995"
)
}
\keyword{datasets}
