% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_0}
\alias{lm_0}
\title{lm_0}
\format{An object of class \code{lm} of length 11.}
\usage{
lm_0
}
\description{
Model of class lm
}
\details{
\code{lm(formula = mpg ~ 1, data = mtcars)}
}
\keyword{datasets}
