% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_2}
\alias{lm_2}
\title{lm_2}
\format{
An object of class \code{lm} of length 12.
}
\usage{
lm_2
}
\description{
Model of class lm
}
\examples{
lm_1 <- lm(mpg ~ wt + cyl, data = mtcars)
}
\keyword{datasets}
