% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_5}
\alias{lm_5}
\title{lm_5}
\format{
An object of class \code{lm} of length 12.
}
\usage{
lm_5
}
\description{
Model of class lm
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
lm(formula = mpg ~ wt + poly(cyl, 2, raw = TRUE), data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
