% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_6}
\alias{lm_6}
\title{lm_6}
\format{An object of class \code{lm} of length 13.}
\usage{
lm_6
}
\description{
Model of class lm
}
\details{
\code{lm(formula = mpg ~ wt * as.factor(gear), data = mtcars)}
}
\keyword{datasets}
