% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_7}
\alias{lm_7}
\title{lm_7}
\format{
An object of class \code{lm} of length 13.
}
\usage{
lm_7
}
\description{
Model of class lm
}
\examples{
lm(formula = mpg ~ as.factor(gear)/wt, data = mtcars)
}
\keyword{datasets}
