% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmerMod_0}
\alias{lmerMod_0}
\title{lmerMod_0}
\format{An object of class \code{lmerMod} of length 1.}
\usage{
lmerMod_0
}
\description{
Model of class lmerMod
}
\examples{
lme4::lmer(wt ~ 1 + (1 | gear), data = mtcars)
}
\keyword{datasets}
