% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmerMod_2}
\alias{lmerMod_2}
\title{lmerMod_2}
\format{
An object of class \code{lmerMod} of length 1.
}
\usage{
lmerMod_2
}
\description{
Model of class lmerMod
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
lme4::lmer(wt ~ drat + cyl + (1 | gear), data = mtcars)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
