% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmerMod_5}
\alias{lmerMod_5}
\title{lmerMod_5}
\format{
An object of class \code{lmerMod} of length 1.
}
\usage{
lmerMod_5
}
\description{
Model of class lmerMod
}
\examples{
data <- iris
data$Cat1 <- rep(c("X", "X", "Y"), length.out = nrow(data))
data$Cat2 <- rep(c("A", "B"), length.out = nrow(data))
lme4::lmer(Petal.Width ~ Cat1 + (1 + Cat1 | Species), data = data)
}
\keyword{datasets}
