% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{methods}
\name{loadAnnotation}
\alias{loadAnnotation}
\alias{loadAnnotation,TxDb-method}
\alias{loadAnnotation,character-method}
\title{Load annotation and prepare a list of features for later}
\usage{
loadAnnotation(txdb)

\S4method{loadAnnotation}{TxDb}(txdb)

\S4method{loadAnnotation}{character}(txdb)
}
\arguments{
\item{txdb}{TxDb gene annotation (or path to SQLite database file)}
}
\value{
A \code{list} of gene feature \code{GRanges}/\code{GRangesList} objects.
}
\description{
Loads a local .sqlite TxDb annotation file (e.g. created using
\code{gtf2sqlite}), returns a list of features needed for circRNA annotation.
}
