% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{loadAnnotation}
\alias{loadAnnotation}
\title{Load annotation and prepare a list of features for later}
\usage{
loadAnnotation(txdb.file)
}
\arguments{
\item{txdb.file}{path to the TxDb gene annotation file saved as SQLite database}
}
\description{
Loads a local .sqlite TxDb annotation file (e.g. created using \code{gtf2sqlite}),
returns a list of features needed for circRNA annotation.
}
