% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merMod_5}
\alias{merMod_5}
\title{merMod_5}
\format{
An object of class \code{glmerMod} of length 1.
}
\usage{
merMod_5
}
\description{
Model of class glmerMod
}
\examples{
\dontrun{
# for data, see '?glmer.nb'
lme4::glmer.nb(y ~ f1 * f2 + (1 | g), data = dd)
}
}
\keyword{datasets}
