% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merMod_nb_1}
\alias{merMod_nb_1}
\title{merMod_nb_1}
\format{
An object of class \code{glmerMod} of length 1.
}
\usage{
merMod_nb_1
}
\description{
Model of class glmerMod
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(lme4)
library(glmmTMB)
data(Owls)
glmer.nb(SiblingNegotiation ~ SexParent + ArrivalTime + (1 | Nest), data = Owls)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
