% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model}
\alias{model}
\title{Model output of random-intercept model from 'brms' using Gryphon data}
\format{
model output with class 'brmsfit', list of 19
\enumerate{
\item BWT:	Birth weight at fledging
\item ANIMAL: Unique identification for each gryphon (1309 levels)
\item MOTHER: Unique identification for each mother (429 levels)
\item BYEAR: Year of Birth
}
}
\usage{
model
}
\description{
This is an example model output from an animal model ran in 'brms' using Gryphon data. This model partitions the variance of birth weight (BWT) into additive genetic variance (ANIMAL), maternal variance (MOTHER) and among birth year variance (BYEAR). The pedigree of individuals (grphyonPed) was included in the model. The model was run with 500 iterations, warmup of 250 and thinned at every sample. Note this is a short run and is for demonstrative purposes only. Estimates will not be reliable.
}
\references{
Wilson, A. J., Reale, D., Clements, M. N., Morrissey, M. M., Postma, E., Walling, C. A., Kruuk, L. E. B & Nussey, D. H. (2010). An ecologist's guide to the animal model. Journal of Animal Ecology, 79(1), 13-26.
}
\seealso{
\code{\link{gryphon}}

\code{\link{gryphonPed}}
}
