% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_bad_3_slopes}
\alias{model_bad_3_slopes}
\title{Model output of a random-slope model from 'brms' using unpublished lizard data}
\format{
model output with class 'brmsfit', list of 19
\enumerate{
\item liz_id:	Unique identification for each lizard
\item dam_id:	Unique identification for each mother
\item id: Unique identification for each lizard, fitting ID of individual twice accounts for permenant environmental variance
\item z_days_since_hatch: days since hatching, z-transformed
}
}
\usage{
model_bad_3_slopes
}
\description{
This is an example model output from an random-slope animal model ran in 'brms' using some unpublished data on body mass of lampropholis delicata collected by Fonti Kar. This model partitions variation in body mass (lnMass) into additive genetic variance, maternal variance (dam_id) as well as permanent environment variance (id). Random slopes was also estimated for maternal and permanent environment variance to model how variance changes with age (z_days_since_hatch). A genomic relatedness matrix was fitted in the model. This model was run with 4000 iterations, warmup of 1500 and thinned at every fifth sample. Note this model is demonstrative purposes only.
}
\references{
Wilson, A. J., Reale, D., Clements, M. N., Morrissey, M. M., Postma, E., Walling, C. A., Kruuk, L. E. B & Nussey, D. H. (2010). An ecologist's guide to the animal model. Journal of Animal Ecology, 79(1), 13-26.
}
\seealso{
\code{\link{gryphon}}

\code{\link{gryphonPed}}
}
