% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_bad_lm}
\alias{model_bad_lm}
\title{Model output of a linear model from 'stats' using Gryphon data}
\format{
model output with class 'lm', list of 13
\enumerate{
\item BWT:	Birth weight at fledging
\item SEX:	Sex of gryphon
}
}
\usage{
model_bad_lm
}
\description{
This is an example model output from an linear model ran in 'stats' using Gryphon data. This model tests for sex differences (SEX) in birth weight (BWT). Note This model is for demonstrative purposes only. Estimates will not be reliable as assumptions of the linear model are not met.
}
\references{
Wilson, A. J., Reale, D., Clements, M. N., Morrissey, M. M., Postma, E., Walling, C. A., Kruuk, L. E. B & Nussey, D. H. (2010). An ecologist's guide to the animal model. Journal of Animal Ecology, 79(1), 13-26.
}
\seealso{
\code{\link{gryphon}}

\code{\link{gryphonPed}}
}
