% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{MungeColumn}
\alias{MungeColumn}
\title{Title}
\usage{
MungeColumn(merge.fos, circ.gr, circ.gr.reduced, column.name)
}
\arguments{
\item{merge.fos}{merge fos}

\item{circ.gr}{circs}

\item{circ.gr.reduced}{reduced circs}

\item{column.name}{column to extract}
}
\value{
a matrix
}
\description{
Function that, based on circRNA candidate list and collapsed circRNA candidate list
summarizes a numeric input column into a matrix that can be hooked to SummarizedExperiment
}
