% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordbetareg_1}
\alias{ordbetareg_1}
\title{ordbetareg_1}
\format{
An object of class \code{brmsfit} (inherits from \code{ordbetareg}) of length 25.
}
\usage{
ordbetareg_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
library(ordbetareg)
data(sleepstudy, package = "lme4")
sleepstudy$y <- datawizard::normalize(sleepstudy$Reaction)
m <- ordbetareg(y ~ Days + (Days | Subject), data = sleepstudy)
}
}
\keyword{datasets}
