% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{partition.coefficient}
\alias{partition.coefficient}
\title{Calculate the partition coefficient}
\usage{
partition.coefficient(clres)
}
\arguments{
\item{clres}{results of a cluster algorithm}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate the partition coefficient
}

