% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{partition.entropy}
\alias{partition.entropy}
\title{Calculate the partition entropy}
\usage{
partition.entropy(clres)
}
\arguments{
\item{clres}{results of a cluster algorithm}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate the partition entropy
}

