% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{ProcessLinear}
\alias{ProcessLinear}
\title{Title}
\usage{
ProcessLinear(dcircs, circ.gr.reduced, wobble)
}
\arguments{
\item{dcircs}{dcircs}

\item{circ.gr.reduced}{reduced circs}

\item{wobble}{how many nucleotides of wobble to tolerate?}
}
\value{
a list
}
\description{
Function that extracts the linear splicing isoforms for each circ RNA
}
