% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{proportion.exponent}
\alias{proportion.exponent}
\title{Calculate the proportion exponent cluster performance metric}
\usage{
proportion.exponent(clres)
}
\arguments{
\item{clres}{results of a cluster algorithm}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate the proportion exponent cluster performance metric
}

