% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{qualFilter}
\alias{qualFilter}
\title{filter find_circ.py-generated splice candidates}
\usage{
qualFilter(sites, n_uniq_thr = 2)
}
\arguments{
\item{sites}{splice site candidates, as read from find_circ.py output by readCircs}

\item{n_uniq_thr}{minimal number of unique reads required for calling a splice site

filter find_circ.py-generated splice candidates

description

details}

\item{sites}{splice site candidates, as read from find_circ.py output by readCircs}

\item{n_uniq_thr}{minimal number of unique reads required for calling a splice site}
}
\description{
description
}
\details{
details
}

