% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\docType{methods}
\name{resTable}
\alias{resTable}
\alias{resTable,RangedSummarizedExperiment-method}
\title{resTable}
\usage{
resTable(se, ...)

\S4method{resTable}{RangedSummarizedExperiment}(se, ...)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{...}{other arguments}
}
\value{
returns a data.table
}
\description{
Function that reads multiple find_circ output files and returns a
SummarizedExperiment object
}
