% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{separation.index}
\alias{separation.index}
\title{Calculate the separation index cluster performance metric}
\usage{
separation.index(clres, x)
}
\arguments{
\item{clres}{results of a cluster algorithm}

\item{x}{a dataset that the cluster algorithm was fit on}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate the separation index cluster performance metric
}

