% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stan_prior_checks_1}
\alias{stan_prior_checks_1}
\title{stan_prior_checks_1}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 28.
}
\usage{
stan_prior_checks_1
}
\description{
Model of class stanreg
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontrun{
library(rstanarm)
prob_fall <- qlogis(0.05)
prob_fall_scale <- 0.5
prior_intercept <- student_t(
  df = 5,
  location = prob_fall,
  scale = prob_fall_scale
)
# believer prior
prob_dem_mid <- log(2)
prob_dem_hi <- log(3.5)

m_believer_ppc <- stan_glm(
  fall_incidence ~ stay + age + mmse + sex,
  data = d,
  family = binomial("logit"),
  prior = student_t(
    df = 5,
    location = c(0, 0, prob_dem_mid, prob_dem_hi, 0, 0, 0, 0, 0, 0, 0),
    scale = c(1, 1, 0.5, 0.5, 1, 1, 1, 1, 1, 1, 1)
  ),
  prior_intercept = prior_intercept,
  refresh = 0,
  cores = 4,
  seed = 1207,
  # only sample from the prior predictive distribution
  prior_PD = TRUE,
  open_progress = FALSE
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
