% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stan_prior_checks_2}
\alias{stan_prior_checks_2}
\title{stan_prior_checks_2}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 28.
}
\usage{
stan_prior_checks_2
}
\description{
Model of class stanreg
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontrun{
library(rstanarm)
# weakly informative prior
stan_glm(
  fall_incidence ~ stay + age + mmse + sex,
  data = d,
  family = binomial("logit"),
  refresh = 0,
  cores = 4,
  seed = 1207,
  # only sample from the prior predictive distribution
  prior_PD = TRUE,
  open_progress = FALSE
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
