% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanmvreg_1}
\alias{stanmvreg_1}
\title{stanmvreg_1}
\format{An object of class \code{stanmvreg} (inherits from \code{stanreg}, \code{lmerMod}) of length 20.}
\usage{
stanmvreg_1
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
library(rstanarm)
stan_mvmer(
  formula = list(
    logBili ~ year + (1 | id),
    albumin ~ sex + year + (year | id)
  ),
  data = pbcLong,
  # this next line is only to keep the example small in size!
  chains = 1, cores = 1, seed = 12345, iter = 1000
)
}
}
\keyword{datasets}
