% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanmvreg_2}
\alias{stanmvreg_2}
\title{stanmvreg_2}
\format{
An object of class \code{stanmvreg} (inherits from \code{stanreg}, \code{lmerMod}) of length 20.
}
\usage{
stanmvreg_2
}
\description{
Model of class stanmvreg
}
\examples{
\dontrun{
library(rstanarm)
library(mediation)
data(jobs)
stan_mvmer(
  list(job_seek ~ treat + econ_hard + sex + age + (1 | occp),
       depress2 ~ treat + job_seek + econ_hard + sex + age + (1 | occp)),
  data = jobs,
  cores = 4,
  seed = 1234,
  refresh = 0)
}
}
\keyword{datasets}
