% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_bernoulli_1}
\alias{stanreg_bernoulli_1}
\title{stanreg_bernoulli_1}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 28.
}
\usage{
stanreg_bernoulli_1
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
set.seed(1234)
dat <- data.frame(
  outcome = rbinom(n = 100, size = 1, prob = 0.35),
  var_binom = as.factor(rbinom(n = 100, size = 1, prob = 0.3)),
  var_cont = rnorm(n = 100, mean = 10, sd = 7),
  groups = sample(letters[1:2], size = 100, replace = TRUE)
)
set.seed(1234)
stan_glm(outcome ~ var_binom * groups + var_cont, data = dat, family = binomial())
}
}
\keyword{datasets}
