% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_fullrank_lm_1}
\alias{stanreg_fullrank_lm_1}
\title{stanreg_fullrank_lm_1}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_fullrank_lm_1
}
\description{
Model of class stanreg
}
\keyword{datasets}
