% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_gam_1}
\alias{stanreg_gam_1}
\title{stanreg_gam_1}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{gamm4}) of length 27.}
\usage{
stanreg_gam_1
}
\description{
Model of class stanreg
}
\keyword{datasets}
