% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_glm_0}
\alias{stanreg_glm_0}
\title{stanreg_glm_0}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 26.}
\usage{
stanreg_glm_0
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
rstanarm::stan_glm(vs ~ 1, data = mtcars, family = "binomial")
}
}
\keyword{datasets}
