% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_glm_2}
\alias{stanreg_glm_2}
\title{stanreg_glm_2}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_glm_2
}
\description{
Model of class stanreg
\dontrun{
rstanarm::stan_glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
}
}
\keyword{datasets}
