% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_glm_4}
\alias{stanreg_glm_4}
\title{stanreg_glm_4}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.
}
\usage{
stanreg_glm_4
}
\description{
Model of class stanreg
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
rstanarm::stan_glm(vs ~ wt + cyl, data = mtcars, family = binomial(link = "probit"))
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
