% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_glm_6}
\alias{stanreg_glm_6}
\title{stanreg_glm_6}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_glm_6
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
set.seed(123)
rstanarm::stan_glm(Sepal.Width ~ Species * Petal.Length, data = iris)
}
}
\keyword{datasets}
