% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_1}
\alias{stanreg_lm_1}
\title{stanreg_lm_1}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.
}
\usage{
stanreg_lm_1
}
\description{
Model of class stanreg
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
rstanarm::stan_glm(mpg ~ wt, data = mtcars)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
