% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_2}
\alias{stanreg_lm_2}
\title{stanreg_lm_2}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_lm_2
}
\description{
Model of class stanreg
}
\examples{
rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
}
\keyword{datasets}
