% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_5}
\alias{stanreg_lm_5}
\title{stanreg_lm_5}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_lm_5
}
\description{
Model of class stanreg
}
\details{
\code{rstanarm::stan_glm(mpg ~ wt + poly(cyl, 2, raw = TRUE), data = mtcars)}
}
\keyword{datasets}
