% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_6}
\alias{stanreg_lm_6}
\title{stanreg_lm_6}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_lm_6
}
\description{
Model of class stanreg
}
\keyword{datasets}
