% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_7}
\alias{stanreg_lm_7}
\title{stanreg_lm_7}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.}
\usage{
stanreg_lm_7
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
rstanarm::stan_glm(mpg ~ as.factor(gear)/wt, data = mtcars)
}
}
\keyword{datasets}
