% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lmerMod_0}
\alias{stanreg_lmerMod_0}
\title{stanreg_lmerMod_0}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{lmerMod}) of length 27.
}
\usage{
stanreg_lmerMod_0
}
\description{
Model of class stanreg
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
rstanarm::stan_lmer(wt ~ 1 + (1 | gear), data = mtcars)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
