% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_merMod_2}
\alias{stanreg_merMod_2}
\title{stanreg_merMod_2}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{lmerMod}) of length 27.}
\usage{
stanreg_merMod_2
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
stan_glmer(vs ~ drat + cyl + (1 | gear), data = mtcars, family = "binomial")
}
}
\keyword{datasets}
