% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_merMod_4}
\alias{stanreg_merMod_4}
\title{stanreg_merMod_4}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{lmerMod}) of length 27.}
\usage{
stanreg_merMod_4
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
stan_glmer(
  vs ~ cyl + (1 | gear),
  data = mtcars,
  family = binomial(link = "probit")
)
}
}
\keyword{datasets}
