% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_merMod_5}
\alias{stanreg_merMod_5}
\title{stanreg_merMod_5}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{lmerMod}) of length 27.
}
\usage{
stanreg_merMod_5
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
stan_glmer(
  cbind(incidence, size - incidence) ~ size + period + (1 | herd),
  data = lme4::cbpp, family = binomial, QR = TRUE,
  chains = 2, cores = 1, seed = 12345, iter = 500, refresh = 0
)
}
}
\keyword{datasets}
