% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_merMod_6}
\alias{stanreg_merMod_6}
\title{stanreg_merMod_6}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}, \code{lmerMod}) of length 28.
}
\usage{
stanreg_merMod_6
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
stan_glmer(
  Reaction ~ Days + (1 + Days | Subject),
  data = lme4::sleepstudy, chains = 2, cores = 1
  seed = 12345, iter = 500, refresh = 0
)
}
}
\keyword{datasets}
