% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\docType{methods}
\name{summarizeCircs}
\alias{summarizeCircs}
\alias{summarizeCircs,character-method}
\title{Function that reads multiple find_circ output files and returns a
SummarizedExperiment object}
\usage{
summarizeCircs(circ.files, keep.linear = TRUE, wobble = 5, subs = "all",
  qualfilter = TRUE, keepCols = 1:6, colData = NULL, ...)

\\S4method{summarizeCircs}{character}(files, keep.linear, wobble, subs, qualfilter, keepCols,colData)
}
\arguments{
\item{circ.files}{a character vector of paths to find_circ output files}

\item{keep.linear}{a boolean indicating whether to keep the counts of
linearly spliced transcripts}

\item{wobble}{number of nucleotides around the splicing border that should be
considered when collapsing circular transcripts - helps with mapping
imprecision}

\item{subs}{a character string, keep only lines containing it in the name column}

\item{qualfilter}{a boolean, tells whether the quality filtering should be performed}

\item{keepCols}{a vector of column numbers return}

\item{colData}{a \code{DataFrame} object that contains the experiment data. It has
to have the same number of rows as the number of files}
}
\value{
returns a \code{SummarizedExperiment}
}
\description{
Function that reads multiple find_circ output files and returns a
SummarizedExperiment object
}
\examples{


}

