% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testCoordinateIndexing}
\alias{testCoordinateIndexing}
\title{test if input and annotation have the same coordinate indexing}
\usage{
testCoordinateIndexing(g1, g2)
}
\arguments{
\item{g1}{a GenomicRanges object}

\item{g2}{a GenomicRanges object}
}
\value{
a list
}
\description{
test if input and annotation have the same coordinate indexing
}
