% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\docType{methods}
\name{uniqReadsQC}
\alias{uniqReadsQC}
\alias{uniqReadsQC,RangedSummarizedExperiment-method}
\title{A scatterplot of total to unique reads supporting a head-to-tail junction}
\usage{
uniqReadsQC(se, sample)

\S4method{uniqReadsQC}{RangedSummarizedExperiment}(se, sample)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{sample}{Which sample to plot?}
}
\value{
ggplot2 scatterplot
}
\description{
Plots a total number of head-to-tail junction reads to log2-fold-change of
total to unique reads ratio. Provies an insight into PCR duplication events.
}
