% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vgam_2}
\alias{vgam_2}
\title{vgam_2}
\format{
An object of class \code{vgam} of length 1.
}
\usage{
vgam_2
}
\description{
Model of class vgam
}
\examples{
data("hunua")
vgam(
  cbind(agaaus, kniexc) ~ vitluc + s(altitude, df = c(2, 3)),
  binomialff(multiple.responses = TRUE),
  data = hunua
)
}
\keyword{datasets}
