% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\docType{methods}
\name{histogram}
\alias{histogram}
\title{histogram}
\usage{
histogram(se, binwidth = 0.7, ...)
}
\value{
returns a ggplot2 object
}
\description{
histogram of SummarizedEXperiment read counts
}

