% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\docType{methods}
\name{writeBedTracks}
\alias{writeBedTracks}
\alias{writeBedTracks,GRangesList-method}
\alias{writeBedTracks,RangedSummarizedExperiment-method}
\title{writeBedTracks}
\usage{
writeBedTracks(circs, ...)

\S4method{writeBedTracks}{GRangesList}(circs, out.prefix = "ciRcus_",
  out.suffix = ".bed", seqlevels.style = "UCSC", sort.bed = TRUE)

\S4method{writeBedTracks}{RangedSummarizedExperiment}(circs, ...)
}
\arguments{
\item{circs}{\code{GRangesList} or \code{SummarizedExperiment} object with
circRNA information}

\item{...}{named arguments defined above to be passed on from
\{code{SummarizedExperiment}-method to \code{GRangesList}-method}

\item{out.prefix}{\code{character} vector (only first element will be used)
specifying the prefix to use for output files (before the
sample name)}

\item{out.suffix}{\code{character} vector (only first element will be used)
specifying the suffix to use for output files (after the
sample name)}

\item{seqlevels.style}{\code{character} vector (only first element will be
used) specifying the seqlevels style to use for the
output BED files}

\item{sort.bed}{\code{logical} vector (only first element will be used)
specifying whether or not to sort the \code{BED} output (by
genomic coordinates)}
}
\value{
None
}
\description{
Write circRNA information from given \code{BED6}-style \code{GRangesList} (or
\code{SummarizedExperiment} object) to \code{BED} files (one per sample).
These can be loaded as genome browser tracks showing circRNA candidates.
}
