% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{AnnotateRanges}
\alias{AnnotateRanges}
\title{title}
\usage{
AnnotateRanges(r1, l, ignore.strand = FALSE, type = "precedence",
  null.fact = "None", collapse.char = ":")
}
\arguments{
\item{r1}{a GenomicRanges object with ranges to be annotated}

\item{l}{a GRangesList with genomic features}

\item{ignore.strand}{should strand info be ignored}

\item{type}{annotation mode, \code{precedence} or \code{all}}

\item{null.fact}{a value to be used for ranges that do not hit any annotation}

\item{collapse.char}{a collapse character for multiple hits in \code{all} mode}
}
\description{
description
annotates the ranges with the corresponding list
}
\details{
details
}
