% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_2}
\alias{anova_2}
\title{anova_2}
\format{An object of class \code{anova} (inherits from \code{data.frame}) with 8 rows and 5 columns.}
\usage{
anova_2
}
\description{
Model of class anova
}
\keyword{datasets}
