% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aovlist_2}
\alias{aovlist_2}
\title{aovlist_2}
\format{
An object of class \code{aovlist} (inherits from \code{listof}) of length 3.
}
\usage{
aovlist_2
}
\description{
Model of class aovlist
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- iris
data$Cat1 <- rep(c("X", "X", "Y"), length.out = nrow(data))
data$Cat2 <- rep(c("A", "B"), length.out = nrow(data))

aov(Sepal.Length ~ Species * Cat1 + Error(Cat2), data = data)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
