% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_4}
\alias{brms_4}
\title{brms_4}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_4
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
brm(
  Sepal.Length ~ Sepal.Width * Species,
  data = iris,
  family = gaussian(),
  chains = 2, iter = 2000, warmup = 1000, cores = 2
)}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
