% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_null_1}
\alias{brms_null_1}
\title{brms_null_1}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
brms_null_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
set.seed(333)
brms::brm(mpg ~ 1, data = mtcars)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
