% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_sigma_1}
\alias{brms_sigma_1}
\title{brms_sigma_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_sigma_1
}
\description{
Model of class brmsfit
}
\examples{
\dontrun{
set.seed(123)
brm(bf(mpg ~ hp + (1 | cyl), sigma ~ cyl), data = mtcars, seed = 123)
}
}
\keyword{datasets}
