% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brms_von_mises_1}
\alias{brms_von_mises_1}
\title{brms_von_mises_1}
\format{
An object of class \code{brmsfit} of length 23.
}
\usage{
brms_von_mises_1
}
\description{
Model of class brmsfit
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
df <- data.frame(
  y = runif(200, -pi, pi),
  x = as.factor(rep(c("a", "b"), 100))
)

f <- brms::bf(
  y ~ 0 + Intercept + x,
  kappa ~ 0 + Intercept,
  family = brms::von_mises()
)

brms::brm(
  f,
  data = df,
  refresh = 0,
  algorithm = "pathfinder",
  backend = "cmdstanr"
)
}
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
