% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTdata}
\alias{BTdata}
\title{BTdata}
\format{
A data frame
\enumerate{
\item tarsus:	species Latin name
\item back:	species common name
\item animal:	logaritm of adult mass dimorphism
\item dam:	method of eggs measurments
\item fosternest:	number of eggs
\item hatchdate:	number of clutches
\item sex:	effect size, correlation
}
}
\description{
BTdata
}
