% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{criteriaSummary}
\alias{criteriaSummary}
\title{Calculate cluster performance metrics for fuzzy clusters}
\usage{
criteriaSummary(clres, metric = "Silhouette")
}
\arguments{
\item{clres}{results of a cluster algorithm}

\item{metric}{a character representing the abbreviation of the metric, see
\code{\link{clusterCrit}} for details, "Silhouette" is default}
}
\value{
a numeric value for the cluster performance
}
\description{
Calculate cluster performance metrics for fuzzy clusters
}

