% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{ensg2name}
\alias{ensg2name}
\title{title}
\usage{
ensg2name(ensg, release = "current", organism)
}
\arguments{
\item{ensg}{character vector of ensembl gene ids}
}
\description{
description
}
\details{
details
}

