% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circbase.R
\name{getIDs}
\alias{getIDs}
\title{title}
\usage{
getIDs(circs, organism, assembly, chrom = "chrom", start = "start",
  end = "end", strand = "strand")
}
\arguments{
\item{chrom}{}

\item{start}{}

\item{end}{}
}
\description{
description
}
\details{
details
}

