% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circbase.R
\name{getIDs}
\alias{getIDs}
\title{title}
\usage{
getIDs(circs, organism, assembly, chrom = "chrom", start = "start",
  end = "end", strand = "strand")
}
\arguments{
\item{circs}{circs}

\item{organism}{organism}

\item{assembly}{assembly}

\item{chrom}{chromosome}

\item{start}{start}

\item{end}{end}

\item{strand}{strand}
}
\description{
description
}
\details{
details
}
