% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_0}
\alias{glm_0}
\title{glm_0}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
glm_0
}
\description{
Model of class glm
}
\examples{
glm(formula = vs ~ 1, family = "binomial", data = mtcars)
}
\keyword{datasets}
