% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLMMadaptive_1}
\alias{GLMMadaptive_1}
\title{GLMMadaptive_1}
\format{
An object of class \code{MixMod} of length 21.
}
\usage{
GLMMadaptive_1
}
\description{
Model of class GLMMadaptive
}
\examples{
\dontrun{
data(cbpp, package = "lme4")

GLMMadaptive::mixed_model(
  cbind(incidence, size - incidence) ~ period,
  random = ~ 1 | herd,
  data = cbpp,
  family = binomial
)
}
}
\keyword{datasets}
