% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmmTMB_2}
\alias{glmmTMB_2}
\title{glmmTMB_2}
\format{An object of class \code{glmmTMB} of length 7.}
\usage{
glmmTMB_2
}
\description{
Model of class glmmTMB
}
\examples{
\dontrun{
glmmTMB(count ~ mined + (1 | site),
  ziformula =  ~ mined,
  family = poisson,
  data = Salamanders
)
}
}
\keyword{datasets}
