% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{gtf2sqlite}
\alias{gtf2sqlite}
\title{Build and save a TxDb object containing gene annotation}
\usage{
gtf2sqlite(assembly = c("hg19", "hg38", "mm10", "rn5", "dm6"), db.file)
}
\arguments{
\item{assembly}{abbreviation for one of the supported assemblies}

\item{db.file}{a file to save SQLite database to}
}
\description{
Loads the GTF file for the selected assembly, drops non-standard chromosomes,
adds "chr" prefix to Ensembl chromosome names and saves the SQLite database
to \code{db.file}
}

