% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htest_2}
\alias{htest_2}
\title{htest_2}
\format{An object of class \code{htest} of length 9.}
\usage{
htest_2
}
\description{
Model of class htest
}
\examples{
cor.test(iris$Sepal.Width, iris$Sepal.Length, method = "pearson")
}
\keyword{datasets}
