% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htest_4}
\alias{htest_4}
\title{htest_4}
\format{An object of class \code{htest} of length 9.}
\usage{
htest_4
}
\description{
Model of class htest
}
\examples{
t.test(iris$Sepal.Width, iris$Sepal.Length)
}
\keyword{datasets}
