% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merMod_1}
\alias{merMod_1}
\title{merMod_1}
\format{
An object of class \code{glmerMod} of length 1.
}
\usage{
merMod_1
}
\description{
Model of class merMod
}
\examples{
lme4::glmer(vs ~ cyl + (1 | gear), data = mtcars, family = "binomial")
}
\keyword{datasets}
