% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merMod_2}
\alias{merMod_2}
\title{merMod_2}
\format{An object of class \code{glmerMod} of length 1.}
\usage{
merMod_2
}
\description{
Model of class merMod
}
\examples{
lme4::glmer(vs ~ drat + cyl + (1 | gear), data = mtcars, family = "binomial")
}
\keyword{datasets}
