% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readCircs}
\alias{readCircs}
\title{read a tabular circRNA candidate list}
\usage{
readCircs(file, subs = "all", qualfilter = TRUE, keepCols = 1:6, ...)
}
\arguments{
\item{file}{location of the input file, a character string such as:
"/home/user/find_circ_sites.bed"}

\item{subs}{a character string, keep only lines containing it in the name column}

\item{keepCols}{a vector of column numbers return

read a tabular circRNA candidate list

description

details}

\item{qualFilter}{a boolean, tells whether the quality filtering should be performed}

\item{file}{location of the input file, a character string such as:
"/home/user/find_circ_sites.bed"}

\item{subs}{a character string, keep only lines containing it in the name column}

\item{qualFilter}{a boolean, tells whether the quality filtering should be performed}

\item{keepCols}{a vector of column numbers return}
}
\description{
description
}
\details{
details
}

