% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_glm_5}
\alias{stanreg_glm_5}
\title{stanreg_glm_5}
\format{
An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 27.
}
\usage{
stanreg_glm_5
}
\description{
Model of class stanreg
}
\examples{
\dontrun{
set.seed(123)
mtcars$count <- rpois(nrow(mtcars), 2)
rstanarm::stan_glm(count ~ wt + cyl, data = mtcars, family = "poisson")
}
}
\keyword{datasets}
