% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stanreg_lm_0}
\alias{stanreg_lm_0}
\title{stanreg_lm_0}
\format{An object of class \code{stanreg} (inherits from \code{glm}, \code{lm}) of length 26.}
\usage{
stanreg_lm_0
}
\description{
Model of class stanreg
}
\examples{
stanarm::stan_glm(mpg ~ 1, data = mtcars)
}
\keyword{datasets}
