% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantification.R, R/readData.R, R/summarize.R
\docType{methods}
\name{circLinRatio}
\alias{circLinRatio}
\alias{resTable}
\alias{summarizeCircs}
\alias{summarizeCircs,character-method}
\title{#title
#
#description
#
#details
#
#}
\usage{
circLinRatio(se, ...)

summarizeCircs(circ.files, keep.linear = TRUE, wobble = 5, subs = "all",
  qualfilter = TRUE, keepCols = 1:6, colData = NULL, ...)

\\S4method{summarizeCircs}{character}(files, keep.linear, wobble, subs, qualfilter, keepCols,colData)

resTable(se, ...)
}
\arguments{
\item{circ.files}{a character vector of paths to find_circ output files}

\item{keep.linear}{a boolean indicating whether to keep the counts of
       linearly spliced transcripts}

\item{wobble}{number of nucleotides around the splicing border that should be
considered when collapsing circular transcripts - helps with mapping
imprecision}

\item{subs}{a character string, keep only lines containing it in the name column}

\item{qualfilter}{a boolean, tells whether the quality filtering should be performed}

\item{keepCols}{a vector of column numbers return}

\item{colData}{a \code{DataFrame} object that contains the experiment data. It has
       to have the same number of rows as the number of files}

\item{sites}{splice site candidates, as read from find_circ.py output by readCircs
#}

\item{label.circ}{a string, label used in the name column to distinguish circular from linear splice sites
#}

\item{label.norm}{a string, label used in the name column to distinguish linear from circular splice sites
#
circLinRatio}

\item{bla}{}
}
\value{
returns a data.table

returns a \code{SummarizedExperiment}

returns a data.table
}
\description{
#title
#
#description
#
#details
#
#

Function that reads multiple find_circ output files and returns a
SummarizedExperiment object

Function that reads multiple find_circ output files and returns a
SummarizedExperiment object
}

