% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\docType{methods}
\name{summarizeCircs}
\alias{summarizeCircs}
\alias{summarizeCircs,data.frame-method}
\alias{summarizeCircs,character-method}
\title{load circRNA detection output to a SummarizedExperiment object}
\usage{
summarizeCircs(colData = NULL, keep.linear = TRUE, wobble = 5,
  subs = "all", qualfilter = TRUE, keepCols = 1:6, ...)

\S4method{summarizeCircs}{data.frame}(colData, keep.linear, wobble, subs,
  qualfilter, keepCols)

\S4method{summarizeCircs}{character}(colData, keep.linear, wobble, subs,
  qualfilter, keepCols)
}
\arguments{
\item{colData}{A \code{DataFrame} object that contains the input files
and sample names to be used for further analysis}

\item{keep.linear}{A boolean indicating whether to keep the counts of
linearly spliced transcripts}

\item{wobble}{Number of nucleotides around the splicing border that should be
considered when collapsing circular transcripts - helps with mapping
imprecision}

\item{subs}{A character string, keep only lines containing it in the name column}

\item{qualfilter}{A boolean. Should quality filtering be performed?}

\item{keepCols}{An integer vector. Which input columns should be returned?}

\item{...}{other arguments}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Function that reads an arbitrary number of circRNA lists and returns a
SummarizedExperiment object. The resulting object contains unified circ RNA
coordinates as GRanges and counts of back-spliced and linearly spliced reads
for each circRNA in each sample
}
