% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vgam_1}
\alias{vgam_1}
\title{vgam_1}
\format{An object of class \code{vgam} of length 1.}
\usage{
vgam_1
}
\description{
Model of class vgam
}
\examples{
data("hunua")
vgam(agaaus ~ vitluc + s(altitude, df = 2), binomialff, data = hunua)
}
\keyword{datasets}
